+++ src/gdal_exp.cpp
@@ -125,7 +125,7 @@ int get_cache_used() {
 //' @param dataType Character data type name.
 //' (e.g., common data types include Byte, Int16, UInt16, Int32, Float32).
 //' @param options Optional list of format-specific creation options in a
-//' vector of "NAME=VALUE" pairs 
+//' vector of `"NAME=VALUE"` pairs 
 //' (e.g., \code{options = c("COMPRESS=LZW")} to set LZW 
 //' compression during creation of a GTiff file).
 //' The APPEND_SUBDATASET=YES option can be 
@@ -133,7 +133,8 @@ int get_cache_used() {
 //' @returns Logical indicating success (invisible \code{TRUE}).
 //' An error is raised if the operation fails.
 //' @seealso
-//' [`GDALRaster-class`][GDALRaster], [createCopy()], [rasterFromRaster()]
+//' [`GDALRaster-class`][GDALRaster], [bandCopyWholeRaster()],
+//' [createCopy()], [rasterFromRaster()]
 //' @examples
 //' new_file <- paste0(tempdir(), "/", "newdata.tif")
 //' create(format="GTiff", dst_filename=new_file, xsize=143, ysize=107,
@@ -204,7 +205,7 @@ bool create(std::string format, std::string dst_filename,
 //' or more normally FALSE indicating that the copy may adapt as needed for  
 //' the output format.
 //' @param options Optional list of format-specific creation options in a
-//' vector of "NAME=VALUE" pairs 
+//' vector of `"NAME=VALUE"` pairs 
 //' (e.g., \code{options = c("COMPRESS=LZW")} to set \code{LZW}
 //' compression during creation of a GTiff file).
 //' The APPEND_SUBDATASET=YES option can be 
@@ -212,7 +213,8 @@ bool create(std::string format, std::string dst_filename,
 //' @returns Logical indicating success (invisible \code{TRUE}).
 //' An error is raised if the operation fails.
 //' @seealso
-//' [`GDALRaster-class`][GDALRaster], [create()], [rasterFromRaster()]
+//' [`GDALRaster-class`][GDALRaster], [bandCopyWholeRaster()], [create()],
+//' [rasterFromRaster()]
 //' @examples
 //' lcp_file <- system.file("extdata/storm_lake.lcp", package="gdalraster")
 //' tif_file <- paste0(tempdir(), "/", "storml_lndscp.tif")
@@ -1005,4 +1007,91 @@ Rcpp::IntegerMatrix createColorRamp(int start_index,
 	GDALDestroyColorTable(hColTbl);
 	
 	return col_tbl;
+}
+
+
+//' Copy a whole raster band efficiently
+//'
+//' `bandCopyWholeRaster()` copies the complete raster contents of one band to
+//' another similarly configured band. The source and destination bands must
+//' have the same xsize and ysize. The bands do not have to have the same data
+//' type. It implements efficient copying, in particular "chunking" the copy in
+//' substantial blocks. This is a wrapper for `GDALRasterBandCopyWholeRaster()`
+//' in the GDAL API.
+//'
+//' @param src_filename Filename of the source raster.
+//' @param src_band Band number in the source raster to be copied.
+//' @param dst_filename Filename of the destination raster.
+//' @param dst_band Band number in the destination raster to copy into.
+//' @param options Optional list of transfer hints in a vector of `"NAME=VALUE"`
+//' pairs. The currently supported `options` are:
+//'   * `"COMPRESSED=YES"` to force alignment on target dataset block sizes to
+//'   achieve best compression. 
+//'    * `"SKIP_HOLES=YES"` to skip chunks that contain only empty blocks.
+//'    Empty blocks are blocks that are generally not physically present in the
+//'    file, and when read through GDAL, contain only pixels whose value is the
+//'    nodata value when it is set, or whose value is 0 when the nodata value is
+//'    not set. The query is done in an efficient way without reading the actual
+//'    pixel values (if implemented by the raster format driver, otherwise will
+//'    not be skipped).
+//' @returns Logical indicating success (invisible \code{TRUE}).
+//' An error is raised if the operation fails.
+//'
+//' @seealso
+//' [`GDALRaster-class`][GDALRaster], [create()], [createCopy()],
+//' [rasterFromRaster()]
+//'
+//' @examples
+//' ## copy Landsat data from a single-band file to a new multi-band image
+//' b5_file <- system.file("extdata/sr_b5_20200829.tif", package="gdalraster")
+//' dst_file <- paste0(tempdir(), "/", "sr_multi.tif")
+//' rasterFromRaster(b5_file, dst_file, nbands=7, init=0)
+//' opt <- c("COMPRESSED=YES", "SKIP_HOLES=YES")
+//' bandCopyWholeRaster(b5_file, 1, dst_file, 5, options=opt)
+//' ds <- new(GDALRaster, dst_file, read_only=TRUE)
+//' ds$getStatistics(band=5, approx_ok=FALSE, force=TRUE)
+//' ds$close()
+// [[Rcpp::export(invisible = true)]]
+bool bandCopyWholeRaster(std::string src_filename, int src_band,
+		std::string dst_filename, int dst_band,
+		Rcpp::Nullable<Rcpp::CharacterVector> options = R_NilValue) {
+
+	GDALDatasetH hSrcDS = NULL;
+	GDALRasterBandH hSrcBand = NULL;
+	GDALDatasetH hDstDS = NULL;
+	GDALRasterBandH hDstBand = NULL;
+	CPLErr err;
+	
+	hSrcDS = GDALOpenShared(src_filename.c_str(), GA_ReadOnly);
+	if (hSrcDS == NULL)
+		Rcpp::stop("Open source raster failed.");
+	hSrcBand = GDALGetRasterBand(hSrcDS, src_band);
+	
+	hDstDS = GDALOpenShared(dst_filename.c_str(), GA_Update);
+	if (hDstDS == NULL)
+		Rcpp::stop("Open destination raster failed.");
+	hDstBand = GDALGetRasterBand(hDstDS, dst_band);
+	
+	std::vector<char *> opt_list = {NULL};
+	if (options.isNotNull()) {
+		// cast to the underlying type
+		// https://gallery.rcpp.org/articles/optional-null-function-arguments/
+		Rcpp::CharacterVector options_in(options);
+		opt_list.resize(options_in.size() + 1);
+		for (R_xlen_t i = 0; i < options_in.size(); ++i) {
+			opt_list[i] = (char *) (options_in[i]);
+		}
+		opt_list[options_in.size()] = NULL;
+	}
+	
+	err = GDALRasterBandCopyWholeRaster(hSrcBand, hDstBand, opt_list.data(),
+			GDALTermProgressR, NULL);
+	
+	if (err != CE_None)
+		Rcpp::stop("bandCopyWholeRaster() failed.");
+		
+	GDALClose(hSrcDS);
+	GDALClose(hDstDS);
+		
+	return true;
 }