+++ NEWS.md
@@ -1,8 +1,10 @@
 # gdalraster 1.4.1 (dev)
 
-* add `GDALRaster$flushCache()`: flush all write cached data to disk (2023-09-14)
+* add `bandCopyWholeRaster()`: efficiently copy a whole raster band, wrapper for `GDALRasterBandCopyWholeRaster()` in the GDAL API (2023-09-14)
 
-* add `GDALRaster$infoAsJSON()`: returns output of the `gdalinfo` command-line utility as a JSON-formatted string (2023-09-14)
+* add `GDALRaster$flushCache()`: flush all write cached data to disk (2023-09-13)
+
+* add `GDALRaster$infoAsJSON()`: returns output of the `gdalinfo` command-line utility as a JSON-formatted string (2023-09-13)
 
 * add `createColorRamp()`: wrapper for `GDALCreateColorRamp()` in the GDAL Raster API. Automatically creates a ramp from one color to another (2023-09-13)